Added delay and retries for weak API servers
Some servers (wikiHow) are not happy with fast, numerous requests to the API and
prevent this from happening by refusing connections.
This would mostly happen on pagination requests which is not controllable outside pywikiapi

This adds two params on Site to enable and allow customization of circumvention mechanisms:
- `retry_after_conn`: nb of seconds to wait before retrying after a ConnectionError
- `pre_request_delay`: nb of seconds to wait before each individual request.

Using the latter should avoid seeing ConnectionError, leaving the former a safety net.